QuickAdminPanel file update add payroll employee id in employee table

+++ database/migrations/2020_11_20_000027_add_relationship_fields_to_offices_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToOfficesTable extends Migration
     public function up()
     {
         Schema::table('offices', function (Blueprint $table) {
-            $table->unsignedInteger('region_id');
+            $table->unsignedBigInteger('region_id');
             $table->foreign('region_id', 'region_fk_2497055')->references('id')->on('regions');
         });
     }
