+++ database/migrations/2020_11_20_000028_add_relationship_fields_to_users_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('employee_id')->nullable();
+            $table->unsignedBigInteger('employee_id')->nullable();
             $table->foreign('employee_id', 'employee_fk_2587242')->references('id')->on('employees');
         });
     }
